+++ EssentialFeed/EssentialFeed/Feed Cache/FeedCachePolicy.swift
@@ -1,0 +1,21 @@
+//
+//  FeedCachePolicy.swift
+//  EssentialFeed
+//
+//  Created by Nayem, Mufakkharul | Nil | GSSD on 2025/06/19.
+//
+
+import Foundation
+
+final class FeedCachePolicy {
+    private init() { }
+    
+    private static let calendar = Calendar(identifier: .gregorian)
+    
+    private static var maxCacheAgeInDays: Int { 7 }
+    
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else { return false }
+        return date < maxCacheAge
+    }
+}
